comments: Add 'comment' module, move `comment-mgr` calls to `ac-comments` and `dc-comments`.
