+++ _ur_addons/comment/@test-comment.mts
@@ -1,0 +1,32 @@
+/*///////////////////////////////// ABOUT \\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\*\
+
+  Test Comment
+  
+  To run this:
+    cd _ur
+    ur comment@test-comment
+
+\*\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\ * /////////////////////////////////////*/
+
+import * as UR from '@ursys/core';
+import ACCOMMENTS from './ac-comment.ts';
+
+/// CONSTANTS & DECLARATIONS //////////////////////////////////////////////////
+/// - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+const { PR } = UR;
+const LOG = PR('comment', 'TagRed');
+
+/// HELPER FUNCTIONS //////////////////////////////////////////////////////////
+/// - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+
+/// API METHODS ///////////////////////////////////////////////////////////////
+/// - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+ACCOMMENTS.Init();
+
+/// EXPORTS ///////////////////////////////////////////////////////////////////
+/// - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+
+LOG('test-comments loaded');
+LOG.warn('test-comments warning');
+LOG.info('test-comments info');
+LOG.error('test-comments error');
