+++ package.json
@@ -61,8 +61,8 @@
   "dependencies": {
     "@iarna/toml": "^2.2.5",
     "@json-editor/json-editor": "^2.10.1",
-    "@ursys/core": "file:./_ur",
     "@ursys/addons": "file:./_ur_addons",
+    "@ursys/core": "file:./_ur",
     "babel-polyfill": "^6.26.0",
     "bootstrap": "^4.6.2",
     "classnames": "^2.3.2",
@@ -92,7 +92,7 @@
     "stacktrace-js": "^2.0.2",
     "superstruct": "^1.0.3",
     "type-detect": "^4.0.8",
-    "yargs": "^17.7.2",
-    "ws": "~1.1.5"
+    "ws": "~1.1.5",
+    "yargs": "^17.7.2"
   }
 }
